app/requirements.txt
# Core FastAPI and web development
fastapi==0.115.7
uvicorn==0.20.0
pydantic==2.10.6
python-multipart==0.0.5
httpx==0.27.2
aiofiles==24.1.0

# LangChain and related AI/ML dependencies
langchain==0.3.15
langchain-community==0.3.15
langchain-core==0.3.31
langchain-google-genai==2.0.9
langchain-text-splitters==0.3.5
openai==1.52.2
tiktoken==0.8.0
faiss-cpu==1.9.0.post1
chromadb==0.6.3
numpy==1.26.4
pandas==2.2.2

# Database and ORM (if needed)
sqlalchemy==2.0.37
psycopg2-binary==2.9.10
motor==3.6.1

# Utilities and helpers
python-dotenv==1.0.1
tenacity==9.0.0
rich==13.9.4
loguru==0.7.2

# Optional: For async task queues (e.g., Celery)
celery==5.4.0
redis==5.0.3

# Optional: For frontend integration (if needed)
jinja2==3.1.5
markdown==3.3.6